+++ rust/signed_doc/src/validator/rules/mod.rs
@@ -3,7 +3,10 @@
 
 use futures::FutureExt;
 
-use crate::{providers::CatalystSignedDocumentProvider, CatalystSignedDocument};
+use crate::{
+    providers::{CatalystSignedDocumentProvider, VerifyingKeyProvider},
+    CatalystSignedDocument,
+};
 
 mod content_encoding;
 mod content_type;
@@ -12,6 +15,7 @@ mod id;
 mod parameters;
 mod reply;
 mod section;
+mod signature;
 mod signature_kid;
 mod template;
 mod ver;
@@ -23,6 +27,7 @@ pub(crate) use id::IdRule;
 pub(crate) use parameters::ParametersRule;
 pub(crate) use reply::ReplyRule;
 pub(crate) use section::SectionRule;
+pub(crate) use signature::SignatureRule;
 pub(crate) use signature_kid::SignatureKidRule;
 pub(crate) use template::{ContentRule, ContentSchema};
 pub(crate) use ver::VerRule;
@@ -49,6 +54,8 @@ pub(crate) struct Rules {
     pub(crate) parameters: ParametersRule,
     /// `kid` field validation rule
     pub(crate) kid: SignatureKidRule,
+    /// document's signatures validation rule
+    pub(crate) signature: SignatureRule,
 }
 
 impl Rules {
@@ -59,7 +66,7 @@ impl Rules {
         provider: &Provider,
     ) -> anyhow::Result<bool>
     where
-        Provider: CatalystSignedDocumentProvider,
+        Provider: CatalystSignedDocumentProvider + VerifyingKeyProvider,
     {
         let rules = [
             self.id.check(doc, provider).boxed(),
@@ -74,12 +81,15 @@ impl Rules {
             self.kid.check(doc).boxed(),
         ];
 
-        let res = futures::future::join_all(rules)
+        let rules_res = futures::future::join_all(rules)
             .await
             .into_iter()
             .collect::<anyhow::Result<Vec<_>>>()?
             .iter()
             .all(|res| *res);
-        Ok(res)
+
+        let sig_res = self.signature.check(doc, provider).await?;
+
+        Ok(rules_res && sig_res)
     }
 }
